fix nettoyage du code fin de sprint
