+++ src/app/guards/admin.guard.ts
@@ -3,10 +3,7 @@ import { UsersService } from '../partage/services/users.service';
 import { inject } from '@angular/core';
 
 export const adminGuard: CanActivateFn = (route, state) => {
-  const usersService = inject(UsersService);  
-  //const role = usersService.rolesArray.toString();
-  //const toto = usersService.fetchAuthUserData(); 
-  //const isAdmin = usersService.rolesArray.includes('Administrateur');
+  const usersService = inject(UsersService);    
   const isAdmin = usersService.profil.roles!.includes('Administrateur');
 
   if (isAdmin ) {
