+++ src/app/intranet/nutrition/nutrition.component.html
@@ -1,1 +1,1 @@
-
+<p>nutrition works</p>
